Prevent libvirt driver starting on non Linux platform
The libvirt driver reads under /proc and /sys and depends on os.stat to
detect if two directories are on the same device. So the driver is
heavily Linux dependent. This patch makes sure that the nova-compute
service refuse to start with the libvirt driver on a non Linux platform.

Change-Id: I29c5a678efec4fbc4bd7958ebe6d454ae30701cd