Warn when starting services with older than N-1 computes
Nova services only support computes that are not older than
the previous major release. This patch introduces a check in the
service startup that warns at staring the service if too old computes
are detected.

This is a partial backport from master. As we agreed on the Wallaby PTG
[1] we would like to backport the check but we don't want to make it a
hard failure of the service startup as that felt too harsh for a stable
branch. So this patch changes the behavior from hard failure to emitting
a warning in the log.

This commit also contains two bugfixes from master squashed into it to
avoid merging broken code:

* Improve error handling during service level check
  Change-Id: I89cdf3852266ed93a2ac7cd6261fe269932026ac
  cherry picked from commit 3b44275868e08992a36e9163f533d689f27a0119

* Restore retrying the RPC connection to conductor
  Change-Id: Iad0ba1a02868eebc2f43b1ac843fcc5096cd5c47
  cherry picked from commit 433bee58bc8d7d65edb6a0805021e51972e6bed6

  Conflicts:
      nova/service.py due to Ie15ec8299ae52ae8f5334d591ed3944e9585cf71
      being only a partial backport to stable/victoria

[1] https://etherpad.opendev.org/p/nova-wallaby-ptg

Closes-Bug: #1903545
Related-Bug: #1871482
Closes-Bug: #1904181

Change-Id: Ie15ec8299ae52ae8f5334d591ed3944e9585cf71
(cherry picked from commit aa7c6f87699ec1340bd446a7d47e1453847a637f)