QEMU/KVM: accept vmxnet3 NIC
Add the ability to use vmxnet3 NIC on a host using the QEMU/KVM driver.
Qemu supports the driver since version 1.5, and Libvirt can expose it
properly.

This allow the migration of an ESXi VM to QEMU/KVM, this without any driver changes.
vmxnet3 comes with better performance, and lower latency comparing to an emulated
driver like e1000.

Change-Id: I95203af4ded9a4dbecfb7981b400f0fb988f2d47