libvirt: Add vTPM config support
There are two changes to our XML element generation code needed for vTPM
support: the '<secret>' element needs to gain support for the 'vtpm'
usage type, and a wholly new '<tpm>' element [1].

None of this is actually used yet, outside of tests. That will come
later.

Part of blueprint add-emulated-virtual-tpm

[1] https://libvirt.org/formatsecret.html
[2] https://libvirt.org/formatdomain.html#elementsTpm

Change-Id: I8b7d2d45963160bc8acf19f36f7945c7545570b3
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>