libvirt: Re-enable live snapshot for paused instances
In change If6c4dd6890ad6e2d00b186c6a9aa85f507b354e0, we added PAUSED to
a list of power states that aren't valid for live snapshot. This was
necessary to workaround issues with older versions of libvirt and QEMU,
such as those found on Ubuntu 16.04. Our versions of libvirt are now new
enough that this is no longer necessary, so stop doing it.

A confusing (read: wrong) note is removed.

Change-Id: Ife37c2920f2a85eca3401f4b698bc9260277d5a1
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Related-Bug: #1741667