libvirt: Remove workaround for really old QEMU
QEMU < 1.0.3 did not support live snapshots. Bug #1193146 noted that for
this to be used, the version of the QEMU that the instance is running is
important, not the version that it was created with. To test this, it
used a (duplicated) call to 'abort_job' that verified the newer versions
of QEMU was in use. There can't be any instances still in the wild using
this version of QEMU, since those users would have had to update their
OS (the last version of Ubuntu to provide a suitably old QEMU was 12.04
[1], which is very much EOL), meaning a reboot of the host and possible
live-migration of instances to another host, both of which would result
in a newer process.

[1] https://launchpad.net/ubuntu/precise/+package/qemu-kvm

Change-Id: Ic55d2ae49a1ae3aefd986bd1f52c76e022fb8ee1
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>
Related-Bug: #1193146