libvirt: Load and cache volume drivers on-demand
This change completes a long standing TODO in the libvirt driver where
volume drivers had previously been loaded at startup. While quick this
can result in drivers being dropped from the resulting list on hosts
that don't support the underlying connector protocols as checked in
os-brick.

We should instead load and cache these drivers on-demand ensuring the
same instance is reused by repeat callers to _get_volume_driver.

An additional TODO is left in place to make this dict configurable in
the future, allowing drivers to be replaced by different volume driver
implementations. An example being the replacment of the os-brick `iscsi`
volume driver with a native `iscsi` QEMU driver.

Change-Id: If02a263055376442a116d00c18a1599b436f21d2