hardware: refactor the code for CPU policy sanity check
The code setting the 'dedicated_cpus' for the 'dedicated' policy
instance does not belong to the scope of CPU policy sanity check,
moving it out of the sanity check block.

Also removed a note saying 'get_dedicated_cpu_constraint' should
be public, which seems not necessary.

This is a follow-up commit in addressing comments in this topic [1].

[1] https://review.opendev.org/#/c/716267/14/

Part of blueprint use-pcpu-and-vcpu-in-one-instance

Change-Id: I3fdfde36641dccfc9de60c25779e60fb9aabba82
Signed-off-by: Wang Huaqiang <huaqiang.wang@intel.com>