tests: Remove 'test_servers.ServersTestBase'
We want to add placement-specific stuff to some libvirt tests.
Currently, these tests inherit from 'test_servers.ServersTestBase',
but, thanks to earlier refactors, this is now essentially identical
to '_IntegratedTestBase', which could also benefit from the placement
functionality. Kill two birds with one stone by removing the
'ServersTestBase' base class in favor of '_IntegratedTestBase'. The only
significant fallout from doing to is that the default API version for
'_IntegratedTestBase' is 'v2.1' versus 'v2' for 'ServersTestBase', but
this is relatively easy remedy.

Change-Id: I999fc4aaf14cca02e4ed730afb844a043b59530c
Signed-off-by: Stephen Finucane <stephenfin@redhat.com>