test: add some notes and new tests for mixed instance
This patch is about to be a follow-up patch to address the
comments of [1], add some notes in introducing the host CPU
usage for an instance with an 'isolate' CPU thread policy.

Also added some other test cases to ensure mixed instance
can reserve proper emulator CPU for hypervisor.

Part of blueprint use-pcpu-and-vcpu-in-one-instance

[1] https://review.opendev.org/716269

Change-Id: I264686705f47713a185536e9d1faa214fc05dafe
Signed-off-by: Wang Huaqiang <huaqiang.wang@intel.com>