Fix misleading documentation for live_migration_inbound_addr
Testing on our own deployment which has 'live_migration_tunnelled'
enabled has indicated that 'live_migration_inbound_addr' still
works in this case, despite the documentation suggesting otherwise.

Looking at the code I am lead to believe that
'live_migration_inbound_addr' is set on the migration target, and
is then used by the migration source to replace the '%s' in the
migration URI. Whilst the 'live_migration_uri' is ignored if both
'live_migration_tunnelled' is enabled and
'live_migration_inbound_addr' is set, the docs suggest further
impact on the inbound_addr which doesn't appear to be true.

This patch attempts to clarify the use of the live migration
parameters.

Change-Id: I5a25786413ede23c72f8ccee1ad12497da7f751c