Correct the check_str and pass actual target in FIP pools policy
FIP pools are not associated with any projects and API
controller also does not take any project_id in request
so anyone requesting that API can get the FIP pools list.
current check_str 'admin_or_owner' is not correct for this
polic, it should be RULE_ANY.

Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit corrects the check_str and pass the actual target for
FIP pools policies which is empty dict.

Partial implement blueprint policy-defaults-refresh-deprecated-apis

Change-Id: I06dde5960fdae5ee4e3ce902482361909ac397d6