Delete non tag related device metadata during detach
When the blueprint virt-device-tagged-attach-detach [1] was
implemented the metadata cleanup during interface detach made an
assumption that metadata only exists if the detached vif has tags.
However metadata exists for vif with vlan information as well. In such
case the metadata is leaked during interface detach. This patch removes
this assumption from the detach code path.

[1] https://review.opendev.org/#/c/415520

Closes-Bug: #1887799
Change-Id: Iefdf9c8b024db04867bd5f8794863f0cd65de807