Pass the actual target in FIP policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for networks policies
which is context.project_id itself as nova cannot verify the owner of
FIP. Neutron will return the authorise error is requester is not
owner of FIP.

Partial implement blueprint policy-defaults-refresh-deprecated-apis

Change-Id: I13d37d36bdc294960306271204c83bbdebaeb7b7