Pass the actual target in security_groups policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for networks policies
which is context.project_id itself as nova cannot verify the owner of
security_groups. Neutron will return the authorise error is requester
is not owner of security_group.

Partial implement blueprint policy-defaults-refresh-deprecated-apis

Change-Id: I1ce8ad8a16bddb3f7520a3b4e323b75626928186