Pass the actual target in volumes policy
Currently if target is not passed in context.can(),
it use defauls target which is context.user_id, context.project_id.
These defaults target are not useful as it pass the
context's user_id and project_id only which means we tell
oslo policy to verify the context data with context data.

This commit pass the actual target for networks policies
which is context.project_id itself as nova cannot verify the owner of
volumes. Neutron will return the authorise error is requester is not
owner of volumes.

Partial implement blueprint policy-defaults-refresh-deprecated-apis

Change-Id: I4108f4b98c57174ca4e71dcf53cd0617093d3f88