Removed the host FQDN from the exception message
Deletion of an instance after disabling the hypervisor by a non-admin
 user leaks the host fqdn in fault msg of instance.Removing the
'host' field from the error message of HypervisorUnavaiable
cause it's leaking host fqdn to non-admin users. The admin user will
see the Hypervisor unavailable exception msg but will be able to figure
on which compute host the guest is on and that the connection is broken.

Conflicts:
    nova/virt/libvirt/host.py

NOTE(melwitt): The conflict is because change
I404ca316b1bf2a963106cd34e927934befbd9b12 is not in Ussuri.

Change-Id: I0eae19399670f59c17c9a1a24e1bfcbf1b514e7b
Closes-Bug: #1851587
(cherry picked from commit a89ffab83261060bbb9dedb2b8de6297b2d07efd)