Only unplug vif after the device is detached from libvirt
There is a potential race between unplugging a vif and removing the
related device from the domain XML. During macvtap (hw_veb) unplug nova
tries to reset the MAC and VLAN config of the VF used for the macvtap.
However as it is done before the macvtap device is removed from the
domain, libvirt still feels responsible to the VF too and restores the
configuration. This leads to a situation where the macvtap is removed
successfully but the VF behind it left configured with the MAC address
and VLAN id of the neutron port.

This patch changes the order of the operations. First the device is
detached from libvirt and then the vif is unplugged.

Part of blueprint sriov-interface-attach-detach

Change-Id: Iea126857725502dc2eef6e53894d8755d0e2e7f4