Add lsscsi to bindep
lsscsi is needed for some of the SCSI device identification. This
identification is actually done in os-brick, but due to the way devstack
and potentially other things install from source, os-brick's bindep.txt
file may not be used.

Since Nova needs the package in order for os-brick to operate, we can
just add it to nova so we can make sure it is available.

Change-Id: Iba5307a6c644ae0fe5eb754f1b1e5dbdfc541cb9
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>