Remove dead volume driver code
This cleans up some dead code for handling Cinder volumes for drivers
that have been removed. The impacted drivers are:

Removed in Ocata
* AOE (CORAID)

Removed in Stein
* DRDB
* HGST
* ITRI DISCO

Removed in Train
* Veritas Hyperscale

Removed in Ussuri
* Sheepdog - the project has been defunct for several
  years now and was considered dead code in Cinder

Since these drivers are not useful without Cinder, they are considered
dead code and therefore do not need to go through a deprecation period
in Nova.

Change-Id: Id1c7d2ca8c93e35b71bf9d48668f1b10a947b0ba
Signed-off-by: Sean McGinnis <sean.mcginnis@gmail.com>