db: fix database migrations when name includes dash
If the name of the database used includes a dash, it will fail the
migrations with an error like this:

  You have an error in your SQL syntax; check the manual that corresponds to
  your MySQL server version for the right syntax to use near '-cell0 DEFAULT
  CHARACTER SET utf8' at line 1

Even more difficult, any future migration fails with the instances table
already existing making it very hard to troubleshoot.  This patch fixes
it by wrapping the schema name in the query.

Closes-Bug: #1892370
Change-Id: Ic96dc415396985737b5bdeffed61077af5f245e6